Replace old tests by Mockito tests
